Build out the delete functionality
Create a function for searching through the clientTree json object and removing the deleted client, then update the ui to match.  Once the delete method has been brought into this branch, we will need to hook up the AJAX call inside the removieClientNode function and only perform the recursive search if it successfully deletes the client.