Make the clients that can't be managed look different
Add a disabled class to the client nodes that can not be managed by the current user to visually show which client they can and can't manage.  They will still be able to view the information for the "disabled" clients, but they won't be able to edit or delete it.