Handle profit centers that the user doesn't have a claim to
Users can be client admins and not have access to that clients profit center, so we need to handle instances where that happens.  This will add a temporary option to the profit center drop down for the existing profit center of the client, but the temporary profit centers will be cleared out each time the form is populated to make sure that they don't have that option for their other clients.  There is also validation on the back end to make sure they're not assigned clients to profit centers that they don't have access to.