Transition to new css system
Atomic was working really well for getting stuff laid out quickly, but it was becoming difficult to implement changes when I needed to make changes in multiple places.  This new approach will use atomic flex classes, but more traditional css in other areas, with as much reuseable content as possible.