Fix pendingChanges()
pendingChanges() was selecting an extraneous input element for each selectized input. This caused it to return true even when there were no real changes. Furthermore, pendingChanges() now returns the number of items that have been changed instead of a boolean.

pendingChanges() does not work correctly when editing client details if any item's value is not set. This is a non-issue because this behavior will not be allowed on release.