Modify import style
Three lines for a single import is a bit overkill. Also, many multi-line imports will be gone at some point in the future due to the use of React and further migration of old JavaScript to TypeScript, so the benefits of one imported class or function per line are diminished.