Get back button to actually work properly in Authorized Content
This is tangentially related to the navbar, so I'm putting it in here.  This make the hidden dom elements actually be shown again on the back button.  This also allows us to keep the content cards loaded beneath the content iframe so that the cards don't need to load again when you navigate back.