Make content authorization consistent; improve readibility
Content authorization now requires the accessing user to be a member of the respective selection group and for neither the root content item nor the selection group to be suspended. Having a role of content user is no longer required, as users are now removed from selection groups when that role is unassigned.