Fix the issue with the back button after password reset
A user was experiencing an issue where she would reset her password, and then rather than click the provided "Click here to log in" link, or the "OK" button that would have taken her to the login screen, it appears that she was clicking the back button.  This was causing the password reset form to laod again, but since the token was invalid now (due to the password having been updated), she was receiving a message that the token was invalid and she was sent another password reset email.  The password reset itself was working, but it looked weird to her.  This fix should take the user to the login page now if they try to navigate back to the password reset form.