Change the way the browseRef is used
Now, the element reference to the file upload element is passed into the file-drop-upload component and then clicked by other elements to trigger a file upload.  This is different than previously, where a reference to an external button was passed into the file-drop-upload component.  This leads to fewer issues and a more reliable system.