Add check to show disclaimer on first open if
If the content item has the ContentDisclaimerAlwaysShown flag, then it uses a timestamp stored in the HTTPContext.Session to determine if the disclaimer should be shown. The timestamp prevents an infinite loop of accepting the disclaimer sending the user back to the disclaimer. The timestamp is set 2 seconds in the future from the time of accepting.