Allow submission but return an error message
Not allowing submissions works but could be confusing for users. This will allow the user to see the warning sent by the backend and keep the form in edit mode as well. Also helps make sure we cover the case of attempting to remove a domain while users are still associated with it, instead of only covering a user removing themselves.